Mejorado del codigo en general
El codigo se mejoro en todos los aspectos: includes, orden de espacios, advertencias debido a comparaciones entre datos de distinto tipo e inclusion de comentarios.

+++ Wonder_boy/src/animation/animation.cpp
@@ -18,7 +18,7 @@ namespace the_wonder_boy
 	}
 	Animation::~Animation()
 	{
-		for (int i = 0; i < frameVector.size(); i++)
+		for (int i = 0; i < static_cast<int>(frameVector.size()); i++)
 		{
 			delete frameVector[i];
 		}
@@ -26,9 +26,10 @@ namespace the_wonder_boy
 		cout << "La animacion ha sido eliminada de la memoria.\n\n";
 	}
 
+	// Funciones públicas.
 	void Animation::addFrame(Frame* frame)
 	{
-		frameVector.push_back(frame); // Lo agrega al final.
+		frameVector.push_back(frame); // Agrega un frame al final del vector.
 	}
 	void Animation::update(float elapsed)
 	{
@@ -36,22 +37,22 @@ namespace the_wonder_boy
 
 		float p = progress;
 
-		for (int i = 0; i < frameVector.size(); i++)
+		for (int i = 0; i < static_cast<int>(frameVector.size()); i++)
 		{
 			actualFrame = i;
 
 			p -= frameVector[i]->getDuration();
 
-			if (animationMode == ANIMATION_MODE::LOOP && p > 0.0f && i == frameVector.size() - 1)
+			if (animationMode == ANIMATION_MODE::LOOP && p > 0.0f && i == static_cast<int>(frameVector.size()) - 1)
 			{
 				// Vuelve al comienzo de la iteración (reiniciando la cuenta).
-				i = -1;
-				continue;
+				i = -1; // "i" se convierte en -1 porque al llegar a "continue;", se inicia el for habiendo sumado 1 a "i".
+				continue; // Vuelve al comienzo del loop, ignorando todo lo que está después de esta línea.
 			}
 
-			if (p <= 0.0f || i == frameVector.size() - 1)
+			if (p <= 0.0f || i == static_cast<int>(frameVector.size()) - 1) // Si se llegó al final del vector...
 			{
-				// Deja el frame actual.
+				// Se queda congelado en el último frame.
 				target.setTextureRect(frameVector[i]->getRect());
 				break;
 			}
